+++ astropy/time/tests/test_update_leap_seconds.py
@@ -1,4 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+from concurrent.futures import ThreadPoolExecutor
 from datetime import datetime, timedelta
 
 import pytest
@@ -7,7 +8,8 @@ import erfa
 from astropy.utils import iers
 from astropy.utils.exceptions import AstropyWarning
 
-from astropy.time import update_leap_seconds
+import astropy.time.core
+from astropy.time import update_leap_seconds, Time
 
 
 class TestUpdateLeapSeconds:
@@ -81,3 +83,17 @@ class TestUpdateLeapSeconds:
 
         with pytest.warns(iers.IERSStaleWarning):
             update_leap_seconds(['erfa', expired_file])
+
+    def test_init_thread_safety(self, monkeypatch):
+        # Set up expired ERFA leap seconds.
+        expired = self.erfa_ls[self.erfa_ls['year'] < 2017]
+        expired.update_erfa_leap_seconds(initialize_erfa='empty')
+        # Force re-initialization, even if another test already did it
+        monkeypatch.setattr(astropy.time.core, '_LEAP_SECONDS_CHECK',
+                            astropy.time.core._LeapSecondsCheck.NOT_STARTED)
+        workers = 4
+        with ThreadPoolExecutor(max_workers=workers) as executor:
+            futures = [executor.submit(lambda: str(Time('2019-01-01 00:00:00.000').tai))
+                       for i in range(workers)]
+            results = [future.result() for future in futures]
+            assert results == ['2019-01-01 00:00:37.000'] * workers